RavenDB-13110 - ReplicationLoader initial abstraction
